Rename Spring WebMVC common module
