+++ .fossa.yml
@@ -1185,7 +1185,7 @@ targets:
       target: ':instrumentation:spring:spring-webmvc:spring-webmvc-6.0:library'
     - type: gradle
       path: ./
-      target: ':instrumentation:spring:spring-webmvc:spring-webmvc-common:javaagent'
+      target: ':instrumentation:spring:spring-webmvc:spring-webmvc-common-3.1:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:vertx:vertx-http-client:vertx-http-client-3.0:javaagent'
