+++ instrumentation/spring/spring-webmvc/spring-webmvc-6.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/webmvc/v6_0/HandlerAdapterInstrumentation.java
@@ -22,7 +22,7 @@ import io.opentelemetry.context.Scope;
 import io.opentelemetry.instrumentation.api.semconv.http.HttpServerRoute;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import io.opentelemetry.javaagent.instrumentation.spring.webmvc.common.IsGrailsHandler;
+import io.opentelemetry.javaagent.instrumentation.spring.webmvc.common.v3_1.IsGrailsHandler;
 import jakarta.servlet.http.HttpServletRequest;
 import javax.annotation.Nullable;
 import net.bytebuddy.asm.Advice;
