+++ settings.gradle.kts
@@ -683,8 +683,8 @@ include(":instrumentation:spring:spring-webmvc:spring-webmvc-3.1:wildfly-testing
 include(":instrumentation:spring:spring-webmvc:spring-webmvc-5.3:library")
 include(":instrumentation:spring:spring-webmvc:spring-webmvc-6.0:javaagent")
 include(":instrumentation:spring:spring-webmvc:spring-webmvc-6.0:library")
-include(":instrumentation:spring:spring-webmvc:spring-webmvc-common:javaagent")
-include(":instrumentation:spring:spring-webmvc:spring-webmvc-common:testing")
+include(":instrumentation:spring:spring-webmvc:spring-webmvc-common-3.1:javaagent")
+include(":instrumentation:spring:spring-webmvc:spring-webmvc-common-3.1:testing")
 include(":instrumentation:spring:spring-ws-2.0:javaagent")
 include(":instrumentation:spring:spring-ws-2.0:testing")
 include(":instrumentation:spring:starters:spring-boot-starter")
